resume-ranker/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart

pymupdf
pytesseract
pillow
nltk

groq
python-dotenv

numpy
scikit-learn
sentence-transformers
torch